==> code/03_easymore_remapping.ipynb <==
# subsetting for lake victoria from larger VIC case study:

```
#!/bin/bash
#SBATCH --account=rpp-kshook
#SBATCH --ntasks=1
#SBATCH --mem-per-cpu=4G
#SBATCH --time=6:00:00    # time (DD-HH:MM)
#SBATCH --job-name=subsetting_VIC
#SBATCH --error=errors1

# load needed modules
module load cdo

# subset the values each may take up to 
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1980.nc mizuRoute_mswep_v1_1979_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1980.nc mizuRoute_mswep_v1_1980_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1981.nc mizuRoute_mswep_v1_1981_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1982.nc mizuRoute_mswep_v1_1982_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1983.nc mizuRoute_mswep_v1_1983_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1984.nc mizuRoute_mswep_v1_1984_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1985.nc mizuRoute_mswep_v1_1985_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1986.nc mizuRoute_mswep_v1_1986_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1987.nc mizuRoute_mswep_v1_1987_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1988.nc mizuRoute_mswep_v1_1988_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1989.nc mizuRoute_mswep_v1_1989_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1990.nc mizuRoute_mswep_v1_1990_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1991.nc mizuRoute_mswep_v1_1991_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1992.nc mizuRoute_mswep_v1_1992_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1993.nc mizuRoute_mswep_v1_1993_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1994.nc mizuRoute_mswep_v1_1994_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1995.nc mizuRoute_mswep_v1_1995_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1996.nc mizuRoute_mswep_v1_1996_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1997.nc mizuRoute_mswep_v1_1997_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1998.nc mizuRoute_mswep_v1_1998_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_1999.nc mizuRoute_mswep_v1_1999_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2000.nc mizuRoute_mswep_v1_2000_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2001.nc mizuRoute_mswep_v1_2001_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2002.nc mizuRoute_mswep_v1_2002_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2003.nc mizuRoute_mswep_v1_2003_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2004.nc mizuRoute_mswep_v1_2004_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2005.nc mizuRoute_mswep_v1_2005_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2006.nc mizuRoute_mswep_v1_2006_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2007.nc mizuRoute_mswep_v1_2007_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2008.nc mizuRoute_mswep_v1_2008_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2009.nc mizuRoute_mswep_v1_2009_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2010.nc mizuRoute_mswep_v1_2010_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2011.nc mizuRoute_mswep_v1_2011_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2012.nc mizuRoute_mswep_v1_2012_subset.nc
cdo sellonlatbox,28,37,-5,2.5 mizuRoute_mswep_v1_2013.nc mizuRoute_mswep_v1_2013_subset.nc
```


# loading EASYMORE
from easymore.easymore import easymore

# initializing EASYMORE object
esmr = easymore()

# specifying EASYMORE objects
# name of the case; the temporary, remapping and remapped file names include case name
esmr.case_name                = 'VIC_Lake_Victoria'              
# temporary path that the EASYMORE generated GIS files and remapped file will be saved
esmr.temp_dir                 = '../temporary/'
# name of target shapefile that the source netcdf files should be remapped to
esmr.target_shp               = '../data/shp/Victoria_cat.shp'
esmr.target_shp_ID            = 'ID' # if not provided easymore give ID according to shape order in shapefile
# name of netCDF file(s); multiple files can be specified with *
esmr.source_nc                = '../mizuRoute/input/mizuRoute_mswep_v1_*_subset.nc' #
# name of variables from source netCDF file(s) to be remapped
esmr.var_names                = ['prec','evapw','runoff']
# name of variable longitude in source netCDF files
esmr.var_lon                  = 'lon'
# name of variable latitude in source netCDF files
esmr.var_lat                  = 'lat'
# name of variable time in source netCDF file; should be always time
esmr.var_time                 = 'time'
# location where the remapped netCDF file will be saved
esmr.output_dir               = '../mizuRoute/input/'
# format of the variables to be saved in remapped files,
# if one format provided it will be expanded to other variables
esmr.format_list              = ['f4']
# fill values of the variables to be saved in remapped files,
# if one value provided it will be expanded to other variables
esmr.fill_value_list          = ['-9999.00']
# if required that the remapped values to be saved as csv as well
esmr.complevel                = 9
# if uncommented EASYMORE will use this and skip GIS tasks
esmr.only_create_remap_csv    = True

# execute EASYMORE
esmr.nc_remapper()

import pandas as pd
import numpy as np
import xarray as xr
import os
import netCDF4 as nc4
import shutil

# merge the remapping and save as a netcdf file
df = pd.DataFrame()
file_to_save = '../data/remapping/'+esmr.case_name +'_remapping.nc'
if not os.path.isdir('../data/remapping/'):
    os.makedirs('../data/remapping/')
df = pd.read_csv(esmr.temp_dir+esmr.case_name +'_remapping.csv')

print(df)
    
# sort and get the case
df = df.sort_values(by=['ID_t']) # sort based on ID_t
case = df['easymore_case'].iloc[0].item()

# river network ID and its frequency in intersection with runoff field grid
RN_id, RN_frequency_in_intersection = np.unique(df['ID_t'], return_counts=True)
print("the dimension of overlap between river network and hydrological units:")
print(len(df))
if len(RN_id) == len(RN_frequency_in_intersection):
    print("the dimension of river network and its frequancy are:")
    print(len(RN_id))
    print(len(RN_frequency_in_intersection))
else:
    sys.exit("the dimension of river network and its frequency are not the same")

# get remapping information from easymore remap file
IDmask  = np.array(df['ID_t']) # the ID of river network
weight  = np.array(df['weight']) # the weight of each hydrological unit in river network
i_index = np.array(df['cols']) # cols for case 1 and 2
j_index = np.array(df['rows']) # rows for case 1 and 2
ID_s    = np.array(df['ID_s']) # ID from unstructure mesh case 3


# write the netcdf file
if os.path.isfile(file_to_save):
    os.remove(file_to_save)

with nc4.Dataset(file_to_save, "w", format="NETCDF4") as ncid: # creating the NetCDF file
    
    # define the dimensions
    dimid_ID  = ncid.createDimension('polyid', len(RN_id))  # limited dimensiton equal the number of hruID
    dimid_RMP = ncid.createDimension('intersect', len(df))   # intersection
    
    # Variables RN_id, RN_frequency_in_intersection
    #
    RN_IDvar   = ncid.createVariable('polyid', 'int',  ('polyid', ), fill_value = -9999, zlib=True, complevel=9)
    RN_IDvar.long_name = 'ID of River Network subbasins'
    RN_IDvar.standard_name = 'ID of River Network subbasins'
    RN_IDvar.units = '1'
    RN_IDvar[:] = RN_id
    #
    RN_FRvar   = ncid.createVariable('frequency', 'int', ('polyid', ), fill_value = -9999, zlib=True, complevel=9)
    RN_FRvar.long_name = 'Frequancy of intersection River Network subbasins with hydrological subbasins'
    RN_FRvar.standard_name = 'Frequancy of intersection River Network subbasins with hydrological subbasins'
    RN_FRvar.units = '1'
    RN_FRvar[:] = RN_frequency_in_intersection
    
    # ID_mask, weight, i_index, j_index, ID_s
    #
    IDmaskvar  = ncid.createVariable('IDmask', 'int',  ('intersect', ), fill_value = -9999, zlib=True, complevel=9)
    IDmaskvar.long_name = 'ID of rive network subbasins'
    IDmaskvar.standard_name = 'ID of rive network subbasins'
    IDmaskvar.units = '1'
    IDmaskvar[:] = IDmask
    #
    weightvar   = ncid.createVariable('weight', 'f8',  ('intersect', ), fill_value = -9999, zlib=True, complevel=9)
    weightvar.long_name = 'weight of each hydrological unit in river network subbasins'
    weightvar.standard_name = 'weight of each hydrological unit in river network subbasins'
    weightvar.units = '1'
    weightvar[:] = weight
    
    if case == 1 or case == 2:
        #
        i_indexvar  = ncid.createVariable('i_index', 'int',  ('intersect', ), fill_value = -9999, zlib=True, complevel=9)
        i_indexvar.long_name = 'cols from the source nc file'
        i_indexvar.standard_name = 'cols from the source nc file'
        i_indexvar.units = '1'
        i_indexvar[:] = i_index + 1
        #
        j_indexvar   = ncid.createVariable('j_index', 'int',  ('intersect', ), fill_value = -9999, zlib=True, complevel=9)
        j_indexvar.long_name = 'rows from the source nc file'
        j_indexvar.standard_name = 'rows from the source nc file'
        j_indexvar.units = '1'
        j_indexvar[:] = j_index + 1
        
    if case == 3:
        #
        ID_HRvar   = ncid.createVariable('ID_HR', 'int',  ('intersect', ), fill_value = -9999, zlib=True, complevel=9)
        ID_HRvar.long_name = 'river network ID'
        ID_HRvar.standard_name = 'river network ID'
        ID_HRvar.units = '1'
        ID_HRvar[:] = ID_s
    
    # general attributes for NetCDF file
    ncid.Conventions = 'CF-1.6'
    ncid.Author = 'The data were written by easymore_codes'
    ncid.License = 'MIT'
    ncid.History = 'Created '
    ncid.Source = 'Case: ; remapped by script from library of Shervan Gharari (https://github.com/ShervanGharari/EASYMORE).'
    
    
# show the remap file
ds = xr.open_dataset(file_to_save)
ds

import shutil
shutil.copy(file_to_save, '../mizuRoute/ancillary_data/'+esmr.case_name +'_remapping.nc')
