==> code/06_sensitivity.ipynb <==
# section 1 load all the necessary modules and packages
import numpy as np
import pandas as pd
from   pyviscous import pyviscous as vs
from   pyviscous import plot as vsplt


# Load input-output data
obj_par = pd.read_csv('../mizuRoute_output_all/obj_par.csv')
obj_par = np.array(obj_par[['HYP_Erate_emr',\
                            'HYP_Qrate_emr',\
                            'scale_factor_Ep_temp',\
                            'scale_factor_P_temp',\
                            '1_NSE',\
                            'RMSE']])
x    = obj_par[:,0:-2]
y    = obj_par[:,-2].reshape(-1,1)

sesitivity = pd.DataFrame(columns=['HYP_Erate_emr',\
                                  'HYP_Qrate_emr',\
                                  'scale_factor_Ep_temp',\
                                  'scale_factor_P_temp'],
                          index = ['first','total'])

for i in [0,1,2,3]:
    sens_indx_first, gmcm_first = vs.viscous(x, y, i, sensType='first', n_components='optimal')  
    sens_indx_total, gmcm_first = vs.viscous(x, y, i, sensType='total', n_components='optimal')
    sesitivity.loc['first'].iloc[i] = sens_indx_first
    sesitivity.loc['total'].iloc[i] = sens_indx_total

sesitivity.to_csv('../mizuRoute_output_all/sesitivity.csv')
sesitivity
